# openapi-schema.yml
openapi: 3.0.3
info:
  title: Remedium Hospital Management System API
  version: 1.0.0
  description: "\nComprehensive REST API for the Remedium Hospital Management System.\n\
    \n## Authentication\nAll API endpoints require authentication. Obtain a token\
    \ via `/api-token-auth/` and include it in requests:\n```\nAuthorization: Token\
    \ <your-token>\n```\n\n## Versioning\nAll API endpoints are versioned under `/api/v1/`.\n\
    \n## Rate Limiting\n- Anonymous users: 100 requests/hour\n- Authenticated users:\
    \ 1000 requests/hour\n- Login attempts: 20/hour\n\n## Resources\n- **Patients**\
    \ - Patient demographics and medical records\n- **Staff** - Staff directory and\
    \ user management\n- **Appointments** - Appointment scheduling and management\n\
    - **Invoices** - Billing and payment tracking\n- **Lab Tests** - Laboratory test\
    \ management\n- **Prescriptions** - Pharmacy and prescription management\n- **Integrations**\
    \ - External system integrations (Admin only)\n    "
  contact:
    name: Remedium HMS Support
    email: support@remediumhms.com
  license:
    name: MIT License
paths:
...
  /api/v1/appointments/scheduled/:
    get:
      operationId: api_v1_appointments_scheduled_retrieve
      description: Get all scheduled appointments.
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
          description: ''
  /api/v1/appointments/upcoming/:
    get:
      operationId: api_v1_appointments_upcoming_retrieve
      description: Get upcoming appointments.
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
          description: ''
...
  /api/v1/invoices/{id}/mark_paid/:
    post:
      operationId: api_v1_invoices_mark_paid_create
      description: Mark an invoice as paid by creating a payment record.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this invoice.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
        required: true
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
          description: ''
...
  /api/v1/invoices/unpaid/:
    get:
      operationId: api_v1_invoices_unpaid_retrieve
      description: Get all unpaid invoices.
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
          description: ''
  /api/v1/lab-tests/:
    get:
      operationId: api_v1_lab_tests_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLabTestList'
          description: ''
    post:
      operationId: api_v1_lab_tests_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabTestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LabTestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LabTestRequest'
        required: true
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTest'
          description: ''
...
  /api/v1/patients/:
    get:
      operationId: api_v1_patients_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPatientFullList'
          description: ''
    post:
      operationId: api_v1_patients_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientFullRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatientFullRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatientFullRequest'
        required: true
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientFull'
          description: ''
...
  /api/v1/patients/{id}/discharge/:
    post:
      operationId: api_v1_patients_discharge_create
      description: Discharge a patient.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this patient.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientFullRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatientFullRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatientFullRequest'
        required: true
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientFull'
          description: ''
...
  /api/v1/prescriptions/:
    get:
      operationId: api_v1_prescriptions_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPrescriptionList'
          description: ''
    post:
      operationId: api_v1_prescriptions_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PrescriptionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PrescriptionRequest'
        required: true
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
          description: ''
...
  /patients/:
    get:
      operationId: patients_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - patients
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPatientList'
          description: ''
components:
  schemas:
    Appointment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        patient:
          type: integer
        patient_detail:
          allOf:
          - $ref: '#/components/schemas/PatientFull'
          readOnly: true
        doctor:
          type: integer
        doctor_detail:
          allOf:
          - $ref: '#/components/schemas/Staff'
          readOnly: true
        appointment_date:
          type: string
          format: date-time
        reason:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/AppointmentStatusEnum'
      required:
      - appointment_date
      - doctor
      - doctor_detail
      - id
      - patient
      - patient_detail
...
    AppointmentStatusEnum:
      enum:
      - Scheduled
      - Completed
      - Cancelled
      type: string
      description: |-
        * `Scheduled` - Scheduled
        * `Completed` - Completed
        * `Cancelled` - Cancelled
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        patient:
          type: integer
        patient_detail:
          allOf:
          - $ref: '#/components/schemas/PatientFull'
          readOnly: true
        issue_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        paid:
          type: boolean
        insurance_claimed:
          type: boolean
        details:
          type: string
          nullable: true
      required:
      - id
      - patient
      - patient_detail
      - total_amount
    InvoiceRequest:
      type: object
      properties:
        patient:
          type: integer
        issue_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        paid:
          type: boolean
        insurance_claimed:
          type: boolean
        details:
          type: string
          nullable: true
      required:
      - patient
      - total_amount
    LabTest:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        test_name:
          type: string
          maxLength: 100
        requested_date:
          type: string
          format: date-time
          readOnly: true
        result_date:
          type: string
          format: date-time
          nullable: true
        result:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/LabTestStatusEnum'
        patient:
          type: integer
      required:
      - id
      - patient
      - requested_date
      - test_name
    LabTestRequest:
      type: object
      properties:
        test_name:
          type: string
          minLength: 1
          maxLength: 100
        result_date:
          type: string
          format: date-time
          nullable: true
        result:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/LabTestStatusEnum'
        patient:
          type: integer
      required:
      - patient
      - test_name
    LabTestStatusEnum:
      enum:
      - Requested
      - Completed
      - Cancelled
      type: string
      description: |-
        * `Requested` - Requested
        * `Completed` - Completed
        * `Cancelled` - Cancelled
...
    PaginatedLabTestList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/LabTest'
    PaginatedPatientFullList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/PatientFull'
    PaginatedPatientList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Patient'
    PaginatedPrescriptionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Prescription'
...
    PatientFull:
      type: object
      description: Clinical staff view - includes all fields including PHI.
      properties:
        id:
          type: integer
          readOnly: true
        unique_id:
          type: string
          description: Unique patient identifier
          maxLength: 20
        first_name:
          type: string
          description: Patient's first name
          maxLength: 50
        last_name:
          type: string
          description: Patient's last name
          maxLength: 50
        full_name:
          type: string
          readOnly: true
        date_of_birth:
          type: string
          format: date
          description: Patient's date of birth
        age:
          type: integer
          readOnly: true
        gender:
          allOf:
          - $ref: '#/components/schemas/PatientGenderEnum'
          description: |-
            Patient's gender

            * `M` - Male
            * `F` - Female
            * `O` - Other
            * `P` - Prefer not to say
        address:
          type: string
          nullable: true
          description: Patient's address
        phone:
          type: string
          nullable: true
          description: Patient's phone number
          pattern: ^\+?\d{9,15}$
          maxLength: 20
        email:
          type: string
          format: email
          nullable: true
          description: Patient's email address
          maxLength: 254
        insurance_provider:
          type: string
          nullable: true
          description: Insurance provider name
          maxLength: 100
        emergency_contact_name:
          type: string
          nullable: true
          description: Emergency contact name
          maxLength: 100
        emergency_contact_phone:
          type: string
          nullable: true
          description: Emergency contact phone number
          pattern: ^\+?\d{9,15}$
          maxLength: 20
        medical_history:
          type: string
          nullable: true
          description: Patient's medical history
        admission_date:
          type: string
          format: date-time
          nullable: true
          description: Hospital admission date and time
        discharge_date:
          type: string
          format: date-time
          nullable: true
          description: Hospital discharge date and time
        is_admitted:
          type: boolean
          readOnly: true
        ward:
          type: integer
          nullable: true
          description: Assigned ward
        room:
          type: integer
          nullable: true
          description: Assigned room
      required:
      - age
      - date_of_birth
      - first_name
      - full_name
      - gender
      - id
      - is_admitted
      - last_name
      - unique_id
    PatientFullRequest:
      type: object
      description: Clinical staff view - includes all fields including PHI.
      properties:
        unique_id:
          type: string
          minLength: 1
          description: Unique patient identifier
          maxLength: 20
        first_name:
          type: string
          minLength: 1
          description: Patient's first name
          maxLength: 50
        last_name:
          type: string
          minLength: 1
          description: Patient's last name
          maxLength: 50
        date_of_birth:
          type: string
          format: date
          description: Patient's date of birth
        gender:
          allOf:
          - $ref: '#/components/schemas/PatientGenderEnum'
          description: |-
            Patient's gender

            * `M` - Male
            * `F` - Female
            * `O` - Other
            * `P` - Prefer not to say
        address:
          type: string
          nullable: true
          description: Patient's address
        phone:
          type: string
          nullable: true
          description: Patient's phone number
          pattern: ^\+?\d{9,15}$
          maxLength: 20
        email:
          type: string
          format: email
          nullable: true
          description: Patient's email address
          maxLength: 254
        insurance_provider:
          type: string
          nullable: true
          description: Insurance provider name
          maxLength: 100
        emergency_contact_name:
          type: string
          nullable: true
          description: Emergency contact name
          maxLength: 100
        emergency_contact_phone:
          type: string
          nullable: true
          description: Emergency contact phone number
          pattern: ^\+?\d{9,15}$
          maxLength: 20
        medical_history:
          type: string
          nullable: true
          description: Patient's medical history
        admission_date:
          type: string
          format: date-time
          nullable: true
          description: Hospital admission date and time
        discharge_date:
          type: string
          format: date-time
          nullable: true
          description: Hospital discharge date and time
        ward:
          type: integer
          nullable: true
          description: Assigned ward
        room:
          type: integer
          nullable: true
          description: Assigned room
      required:
      - date_of_birth
      - first_name
      - gender
      - last_name
      - unique_id
    PatientGenderEnum:
      enum:
      - M
      - F
      - O
      - P
      type: string
      description: |-
        * `M` - Male
        * `F` - Female
        * `O` - Other
        * `P` - Prefer not to say
    Prescription:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        patient:
          type: integer
        drug_name:
          type: string
          maxLength: 100
        dosage:
          type: string
          maxLength: 50
        frequency:
          type: string
          maxLength: 50
        prescribed_date:
          type: string
          format: date-time
          readOnly: true
        prescribed_by:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
      - dosage
      - drug_name
      - frequency
      - id
      - patient
      - prescribed_date
    PrescriptionRequest:
      type: object
      properties:
        patient:
          type: integer
        drug_name:
          type: string
          minLength: 1
          maxLength: 100
        dosage:
          type: string
          minLength: 1
          maxLength: 50
        frequency:
          type: string
          minLength: 1
          maxLength: 50
        prescribed_by:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
      - dosage
      - drug_name
      - frequency
      - patient
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"
servers:
- url: http://localhost:8000
  description: Development server